+++ .github/scripts/x86_64-Linux/pkgs/affine.sh
@@ -87,16 +87,23 @@ if [ "${SKIP_BUILD}" == "NO" ]; then
        export APP="affine"
        export NIX_PKGNAME="affine"
        export PKG_NAME="${NIX_PKGNAME}.NixAppImage"
-       PKG_NIX_DIR="$(find "/opt/nixpkgs/pkgs/by-name" -type d -name "affine" -exec realpath "{}" \;)" && export PKG_NIX_DIR="${PKG_NIX_DIR}"
-       PKG_NIX_TMP="$(find "/opt/nixpkgs/pkgs/by-name" -type d -name "affine" -exec find "{}" -type f -name "package.nix" \;)" && export PKG_NIX_TMP="${PKG_NIX_TMP}"
-       pushd "${PKG_NIX_DIR}" >/dev/null 2>&1
-       curl -qfsSL "https://raw.githubusercontent.com/Azathothas/Toolpacks-Extras/refs/heads/main/.github/assets/nix-flakes/numtide_nix-gl-host.nix" -o "${PKG_NIX_DIR}/flake.nix"
-       sed "s/PKG_NAME/${APP}/g" -i "${PKG_NIX_DIR}/flake.nix"
-       sed "s/PKG_ARCH/$(uname -m)/g" -i "${PKG_NIX_DIR}/flake.nix"
-       git add --all --verbose && git commit -m "[+] NixAppImage ${APP}"
-       git stash ; nix flake update ; popd >/dev/null 2>&1
-       nix bundle --bundler "github:ralismark/nix-appimage" "${PKG_NIX_DIR}" --log-format bar-with-logs
-       pushd "/opt/nixpkgs" >/dev/null 2>&1 && git reset --hard "origin/master" && popd >/dev/null 2>&1
+       #If local repo, use
+       if [ -d "/opt/nixpkgs" ] && [ $(du -s "/opt/nixpkgs" | cut -f1) -gt 100 ]; then
+         pushd "/opt/nixpkgs" >/dev/null 2>&1 && git reset --hard "origin/master" && popd >/dev/null 2>&1
+         PKG_NIX_DIR="$(find "/opt/nixpkgs/pkgs/by-name" -type d -name "affine" -exec realpath "{}" \;)" && export PKG_NIX_DIR="${PKG_NIX_DIR}"
+         PKG_NIX_TMP="$(find "/opt/nixpkgs/pkgs/by-name" -type d -name "affine" -exec find "{}" -type f -name "package.nix" \;)" && export PKG_NIX_TMP="${PKG_NIX_TMP}"
+         pushd "${PKG_NIX_DIR}" >/dev/null 2>&1
+         curl -qfsSL "https://raw.githubusercontent.com/Azathothas/Toolpacks-Extras/refs/heads/main/.github/assets/nix-flakes/numtide_nix-gl-host.nix" -o "${PKG_NIX_DIR}/flake.nix"
+         sed "s/PKG_NAME/${APP}/g" -i "${PKG_NIX_DIR}/flake.nix"
+         sed "s/PKG_ARCH/$(uname -m)/g" -i "${PKG_NIX_DIR}/flake.nix"
+         git add --all --verbose && git commit -m "[+] NixAppImage ${APP}"
+         git stash ; nix flake update ; popd >/dev/null 2>&1
+         nix bundle --bundler "github:ralismark/nix-appimage" "${PKG_NIX_DIR}" --log-format bar-with-logs
+         pushd "/opt/nixpkgs" >/dev/null 2>&1 && git reset --hard "origin/master" && popd >/dev/null 2>&1
+         unset PKG_NIX_DIR PKG_NIX_TMP
+       else
+         nix bundle --bundler "github:ralismark/nix-appimage" "nixpkgs#${NIX_PKGNAME}" --log-format bar-with-logs
+       fi
       #Copy
        sudo rsync -achL "${OWD}/${APP}.AppImage" "${OWD}/${PKG_NAME}.tmp"
        sudo chown -R "$(whoami):$(whoami)" "${OWD}/${PKG_NAME}.tmp" && chmod -R 755 "${OWD}/${PKG_NAME}.tmp"
